moves filters to seperate controller
